Address some review comments.
